utils: Always linkify http/https schemas
When running sandboxed, we might not have access to schema handlers
or the list of installed applications to determine which schemas are
supported. We won't be able to actually open links in that case, but
we assume that we can open URLs without schema (as http://), so it
makes sense to also always linkify http:// URLs with schema. At least
users can copy the URL and paste it into their browser ...